Don't install signal-hook on Windows
