+++ lumen_runtime/src/system/break_handler/wasm32.rs
@@ -1,0 +1,6 @@
+use bus::Bus;
+
+use super::Signal;
+
+// Signal handling doesn't apply to WebAssembly
+pub fn init(_bus: Bus<Signal>) {}
